ADD - Secure level controllers

+++ backend/src/level/level.controller.ts
@@ -2,53 +2,42 @@ import { Controller, Get, Post, Body, Patch, Param, Delete, HttpException, HttpS
 import { LevelService } from './level.service';
 import { CreateLevelDto } from './dto/create-level.dto';
 import { UpdateLevelDto } from './dto/update-level.dto';
+import { Auth } from 'src/utils/decorators/auth.decorator';
+import { Role } from 'src/utils/types/roles.types';
 
 @Controller('level')
 export class LevelController {
   constructor(private readonly levelService: LevelService) {}
 
   @Post()
+  @Auth()
   async create(@Body() createLevelDto: CreateLevelDto) {
-    try {
-      return await this.levelService.create(createLevelDto);
-    } catch (err) {
-      throw new HttpException('Bad request.', HttpStatus.BAD_REQUEST);
-    }
+    return await this.levelService.create(createLevelDto);
   }
 
   @Get()
+  @Auth(Role.STAFF)
   async findAll() {
-    try {
-      return await this.levelService.findAll();
-    } catch (err) {
-      throw new HttpException('Bad request.', HttpStatus.BAD_REQUEST);
-    }
+    return await this.levelService.findAll();
   }
 
   @Get(':id')
+  @Auth()
   async findOne(@Param('id') id: string) {
-    try {
-      return await this.levelService.findOne(id);
-    } catch {
-      throw new HttpException('Bad request.', HttpStatus.BAD_REQUEST);
-    }
+    if (!id) throw new HttpException('Bad request', HttpStatus.BAD_REQUEST);
+    const level = await this.levelService.findOne(id);
+    return level;
   }
 
   @Patch(':id')
   async update(@Param('id') id: string, @Body() updateLevelDto: UpdateLevelDto) {
-    try {
-      return await this.levelService.update(id, updateLevelDto);
-    } catch (err) {
-      throw new HttpException('Bad request.', HttpStatus.BAD_REQUEST);
-    }
+    if (!id) throw new HttpException('Bad request', HttpStatus.BAD_REQUEST);
+    return await this.levelService.update(id, updateLevelDto);
   }
 
   @Delete(':id')
   async remove(@Param('id') id: string) {
-    try {
-      return await this.levelService.remove(id);
-    } catch (err) {
-      throw new HttpException('Bad request.', HttpStatus.BAD_REQUEST);
-    }
+    if (!id) throw new HttpException('Bad request', HttpStatus.BAD_REQUEST);
+    return await this.levelService.remove(id);
   }
 }
